requirements.txt
Django==2.0.7
celery==4.3.0
djangorestframework==3.8.2
mysqlclient==1.3.13
djangorestframework-jwt==1.11.0
drf-dynamic-fields == 0.3.0
drf-friendly-errors==0.13
djangorestframework-filters == 0.10.2
django-crontab==0.7.1
django-redis == 4.9.0
django-filter==2.0.0
django-celery==3.2.2
pyotp == 2.2.6
django-cors-headers == 2.2.0
redis==3.2.1
Pillow==5.3.0
requests==2.20.1
flower==0.9.2
IPy==1.0
PyJWT==1.7.1
pymongo==3.7.2
PyMySQL==0.9.3
pyotp==2.2.6
pycrypto==2.6.1
anyjson==0.3.3


